odoo/config/requirements.txt
boto3
git-aggregator
paramiko
num2words
redis
watchdog
xlwt
zeep
requests
pyyaml=5.1.2